requirements.txt
absl-py==0.12.0
alive-progress==1.6.1
aniso8601==9.0.1
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.4
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
cycler==0.10.0
deepface==0.0.51
Django==2.2.6
djangorestframework==3.10.3
elementpath==2.0.5
filelock==3.0.12
firebase==3.0.1
Flask==2.0.0
Flask-RESTful==0.3.8
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==3.13.0
google-api-core==1.24.1
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-cloud-vision==2.3.1
google-crc32c==1.1.0
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.1
gunicorn==19.6.0
h5py==3.1.0
idna==2.10
iso8601==0.1.13
isort==5.6.4
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.0
jwcrypto==0.8
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lightgbm==3.2.1
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.3.3
mccabe==0.6.1
MouseInfo==0.1.3
mtcnn==0.1.0
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
opt-einsum==3.3.0
paho-mqtt==1.5.1
pandas==1.2.0
Pillow==8.0.1
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
PyGetWindow==0.0.9
pylint==2.6.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.1
python-jwt==3.3.0
PyTweening==1.0.3
pytz==2020.5
PyYAML==5.3.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retina-face==0.0.4
rsa==4.7.2
scikit-learn==0.24.0
scipy==1.5.4
seaborn==0.11.1
serial==0.0.97
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tflearn==0.5.0
threadpoolctl==2.1.0
tk==0.1.0
toml==0.10.2
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==12.0.6
Werkzeug==2.0.0
whitenoise==4.1.4
wrapt==1.12.1
